% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylist.R
\name{shinylist}
\alias{shinylist}
\title{HTML lists with R}
\usage{
shinylist(dat, ordered = FALSE, width = NULL, height = NULL)
}
\arguments{
\item{dat}{the vector of data to be output as HTML list}

\item{oderered}{indicator of whether or not list should be ordered (i.e. numbered) or unordered}
}
\description{
Create an HTML list from R input. Vector input will be parsed by element and output as either ordered or unordered list item(s). Can be used inside of Shiny app with \code{\link{listOuput}} and \code{\link{renderList}}.
}

